Actualización del programa (código final) y módulos.
